15712: PR feedback - renamed typescript files and switch to class bind directive.
