+++ web/src/lib/components/asset-viewer/album-list-item.svelte
@@ -4,7 +4,7 @@
   import { normalizeSearchString } from '$lib/utils/string-utils.js';
   import AlbumListItemDetails from './album-list-item-details.svelte';
   import type { Action } from 'svelte/action';
-  import { SCROLL_PROPERTIES } from '$lib/components/shared-components/add-album-modal/album-modal';
+  import { SCROLL_PROPERTIES } from '$lib/components/shared-components/add-album-modal/add-album-modal';
 
   interface Props {
     album: AlbumResponseDto;
@@ -43,8 +43,9 @@
   type="button"
   onclick={onAlbumClick}
   use:scrollIntoViewIfSelected
-  class="flex w-full gap-4 px-6 py-2 text-left transition-colors hover:bg-gray-200 dark:hover:bg-gray-700 rounded-xl
-        {selected ? 'bg-gray-200 dark:bg-gray-700' : ''}"
+  class="flex w-full gap-4 px-6 py-2 text-left transition-colors hover:bg-gray-200 dark:hover:bg-gray-700 rounded-xl"
+  class:bg-gray-200={selected}
+  class:dark:bg-gray-700={selected}
 >
   <span class="h-12 w-12 shrink-0 rounded-xl bg-slate-300">
     {#if album.albumThumbnailAssetId}
